Renaming components, using named exports and index
