+++ app/javascript/src/Navigation/components/header/Header.jsx
@@ -1,0 +1,13 @@
+import React from 'react'
+import { PageHeader } from '@patternfly/react-core'
+import { Brand, SessionMenu, Toolbar } from 'Navigation/components/header'
+
+const Header = ({logoProps, toolbarProps, docsProps, avatarProps}) => (
+  <PageHeader
+    logo = {<Brand/>}
+    logoProps = {logoProps}
+    toolbar = {<Toolbar toolbarProps={toolbarProps} docsProps={docsProps}/>}
+    avatar = {<SessionMenu {...avatarProps}/>}/>
+)
+
+export { Header }
