+++ app/javascript/src/Navigation/components/header/index.jsx
@@ -1,15 +1,9 @@
-import React from 'react'
-import { PageHeader } from '@patternfly/react-core'
-import SessionMenu from 'Navigation/components/header/SessionMenu'
-import Toolbar from 'Navigation/components/header/Toolbar'
-import Brand from 'Navigation/components/header/Brand'
+import {AccountSettingsMenu} from 'Navigation/components/header/AccountSettingsMenu'
+import {Brand} from 'Navigation/components/header/Brand'
+import {DocumentationMenu} from 'Navigation/components/header/DocumentationMenu'
+import {DocumentationMenuItem} from 'Navigation/components/header/DocumentationMenuItem'
+import {Header} from 'Navigation/components/header/Header'
+import {SessionMenu} from 'Navigation/components/header/SessionMenu'
+import {Toolbar} from 'Navigation/components/header/Toolbar'
 
-const Header = ({logoProps, toolbarProps, docsProps, avatarProps}) => (
-  <PageHeader
-    logo={<Brand/>}
-    logoProps = {logoProps}
-    toolbar = {<Toolbar toolbarProps={toolbarProps} docsProps={docsProps}/>}
-    avatar = {<SessionMenu {...avatarProps}/>}/>
-)
-
-export default Header
+export { AccountSettingsMenu, Brand, DocumentationMenu, DocumentationMenuItem, Header, SessionMenu, Toolbar }
